backend/requirements.txt
# ============================================================================
# Stock Screening Platform - Backend Dependencies
# ============================================================================

# Web Framework
fastapi==0.123.0  # Security: Latest stable, supports starlette 0.49.1
starlette==0.50.0  # Security: Fixed HTTP Range DoS (GHSA-7f5h-v6xp-fcq8, CVE-2025-62727)
uvicorn[standard]==0.38.0
gunicorn==23.0.0  # Security: Fixed request smuggling (GHSA-w3h3-4rj7-4ph4, GHSA-hc5x-x2vx-497g)

# Database
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic==1.17.1
psycopg2-binary==2.9.11
greenlet==3.0.3  # Required for SQLAlchemy asyncio

# Cache
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.4.0  # Security: Fixed algorithm confusion and JWT bomb (PYSEC-2024-232, PYSEC-2024-233)
bcrypt==5.0.0
python-multipart==0.0.20  # Security: Fixed ReDoS vulnerabilities (GHSA-2jv5-9r88-3w3p, GHSA-59g5-xgcq-4qw3)
pydantic[email]==2.12.5
pydantic-settings==2.12.0  # Upgraded: Compatible with pydantic 2.11.7

# Payment Processing
stripe==14.0.1

# Background Tasks
# Note: Celery 5.4.0+ supports redis 5.0+
celery[redis]==5.6.0

# Data Processing
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.7.2
lightgbm==4.6.0
xgboost==3.1.2
optuna==4.6.0
openai==1.12.0
anthropic==0.18.1
tenacity==8.2.3
jinja2==3.1.5
mlflow==3.5.0  # Security: Fixed multiple CVEs (PYSEC-2025-17, PYSEC-2025-52, path traversal, DoS)
tensorflow==2.16.1
keras==3.12.0  # Security: Fixed code execution vulnerabilities (GHSA-48g7-3x6r-xfhp, GHSA-c9rc-mg46-23w3)
matplotlib==3.8.2
pillow==10.4.0  # Security: Fixed buffer overflow (GHSA-44wm-f244-xhp3)
scipy==1.11.4  # Upgraded: Supports gunicorn>=23, fixes dependency conflict

# HTTP Requests
httpx==0.28.1
aiohttp==3.13.2  # Security: Fixed request smuggling in chunked trailer parsing (GHSA-9548-qrrj-x5pj, CVE-2025-53643)

# Monitoring & Logging
prometheus-client==0.23.1
sentry-sdk[fastapi]==2.45.0  # Security: Fixed env variable exposure (GHSA-g92j-qhmh-64v2)

# Testing
pytest==9.0.1
pytest-asyncio==0.25.3
pytest-cov==4.1.0
aiosqlite==0.21.0
freezegun==1.5.5
# httpx already listed in HTTP Requests section

# Code Quality
black==25.11.0  # Security: Fixed ReDoS (PYSEC-2024-48)
flake8==7.3.0
mypy==1.19.0
isort==7.0.0

# Utilities
python-dotenv==1.2.1  # Upgraded: Required by mlflow 3.x (fastmcp dependency)
pyyaml==6.0.1
email-validator==2.1.0.post1
zipp>=3.19.1
tornado>=6.4.1
idna>=3.7
certifi>=2024.07.04
